% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{anchorOrtho}
\alias{anchorOrtho}
\title{Filter ortholog pairs based on Locally Collinear Blocks (LCBs)}
\usage{
anchorOrtho(object, non1to1 = TRUE)

anchorOrtho(object, non1to1 = TRUE)
}
\arguments{
\item{object}{A SynogDB object.}

\item{non1to1}{Logical indicating whether to include non-1-to-1 LCB pairs (default is TRUE).}
}
\description{
This function filters ortholog pairs based on their presence in Locally Collinear Blocks (LCBs).

This function filters ortholog pairs based on their presence in Locally Collinear Blocks (LCBs).
}
