% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{createFASTA}
\alias{createFASTA}
\title{Create FASTA files}
\usage{
createFASTA(object, out_dir)
}
\arguments{
\item{object}{A SynogDB object.}

\item{out_dir}{Output directory for FASTA files.}
}
\description{
This function generates FASTA files for query and subject CDS from the SynogDB object.
}
