% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.addChrLen}
\alias{.addChrLen}
\title{Add chromosome lengths to plot}
\usage{
.addChrLen(p, chrLen)
}
\arguments{
\item{p}{A ggplot object.}

\item{chrLen}{A list containing chromosome lengths.}
}
\value{
A ggplot object with chromosome lengths added.
}
\description{
This function adds chromosome lengths to a ggplot object.
}
