% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.addRBBH}
\alias{.addRBBH}
\title{Add Reciprocal Best BLAST Hits (RBBH) to Syntenic Orthologs}
\usage{
.addRBBH(rbbh, syn_og)
}
\description{
This function identifies candidate Reciprocal Best BLAST Hits (RBBH) that are not already present in the syntenic ortholog list.
}
