% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{.blast_filter}
\alias{.blast_filter}
\title{Filter BLAST search results}
\usage{
.blast_filter(blast_out, pident = 0, qcovs = 0, evalue = Inf, best = FALSE)
}
\arguments{
\item{blast_out}{A data.frame containing the BLAST search results.}

\item{pident}{Minimum percentage identity for BLAST hits (default is 0).}

\item{qcovs}{Minimum query coverage for BLAST hits (default is 0).}

\item{evalue}{Maximum e-value for BLAST hits (default is Inf).}

\item{best}{Logical indicating whether to filter for the best hits (default is FALSE).}
}
\value{
A filtered data.frame of BLAST search results.
}
\description{
This function filters BLAST search results based on specified criteria for percentage identity,
query coverage, and e-value. It can also identify the best hits for each query sequence.
}
