% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{.blast_search}
\alias{.blast_search}
\title{Perform BLAST search}
\usage{
.blast_search(fa, db, blast_path, n_threads)
}
\arguments{
\item{fa}{A DNAStringSet object containing the query sequences.}

\item{db}{A BLAST database connection.}

\item{n_threads}{Number of threads to use for BLAST (default is 1).}

\item{n_batch}{Number of sequences to process in each batch (default is NULL).}

\item{max_target_seqs}{Maximum number of target sequences (default is 100000).}

\item{task}{BLAST task to perform (default is "-task blastn").}
}
\value{
A data.frame containing the BLAST search results.
}
\description{
This function performs a BLAST search using the specified parameters.
}
