% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.calc_lcb_stats}
\alias{.calc_lcb_stats}
\title{Calculate LCB statistics}
\usage{
.calc_lcb_stats(object, target)
}
\arguments{
\item{object}{A data.frame containing LCB information.}

\item{target}{A regex pattern to match the target genome (e.g., "^query_").}
}
\value{
A data.frame with LCB statistics.
}
\description{
This function calculates statistics for a given set of LCBs (Locally Collinear Blocks).
}
