% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.checkLength}
\alias{.checkLength}
\title{Check Length of Transcripts}
\usage{
.checkLength(gff, gff3, len_diff)
}
\description{
This function checks the length of transcripts and filters those that meet the length difference criteria.
}
