% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.classifySO}
\alias{.classifySO}
\title{Classify Orthologs}
\usage{
.classifySO(ortho)
}
\description{
This function classifies ortholog pairs into categories: 1-to-1, 1-to-many, many-to-1, and many-to-many.
}
