% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.classifySOgene}
\alias{.classifySOgene}
\title{Classify Syntenic Ortholog Gene Pairs}
\usage{
.classifySOgene(ortho)
}
\description{
This function classifies syntenic ortholog gene pairs into categories: 1-to-1, 1-to-M, M-to-1, and M-to-M.
}
