% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.condaExe}
\alias{.condaExe}
\title{Execute a command using conda environment}
\usage{
.condaExe(conda, env, command, args)
}
\arguments{
\item{conda}{Path to the conda executable.}

\item{env}{Name of the conda environment to use.}

\item{command}{Command to execute.}

\item{args}{Arguments for the command.}
}
\description{
This function executes a command within a specified conda environment.
}
