% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_orthology.R
\name{.create_gff_lookup_dataframe}
\alias{.create_gff_lookup_dataframe}
\title{Create GFF lookup: genome ID -> GFF file path
RDS can be list(gff_df, cds_df); we use the first element for lookup.}
\usage{
.create_gff_lookup_dataframe(gff_df_paths, genome_width, verbose)
}
\description{
Create GFF lookup: genome ID -> GFF file path
RDS can be list(gff_df, cds_df); we use the first element for lookup.
}
