% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.createFASTA}
\alias{.createFASTA}
\title{Create FASTA files}
\usage{
.createFASTA(in_list, out_dir)
}
\arguments{
\item{out_dir}{Output directory for FASTA files.}

\item{object}{A OrthoPairDB object.}
}
\description{
This function generates FASTA files for query and subject CDS from the OrthoPairDB object.
}
