% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.df2gr}
\alias{.df2gr}
\title{Convert LCB pairs to GRanges objects}
\usage{
.df2gr(pairs)
}
\arguments{
\item{pairs}{A list containing LCB pairs data frames.}
}
\value{
A list of GRanges objects for query and subject genomes.
}
\description{
This function converts Locally Collinear Block (LCB) pairs to GRanges objects for both query and subject genomes.
}
