% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.eval1toM}
\alias{.eval1toM}
\title{Evaluate 1-to-M Orthologs for Splits}
\usage{
.eval1toM(ortho, gff = gff)
}
\description{
This function evaluates 1-to-M orthologs to identify cases where a 1-to-M relationship might actually be a result of gene splits.
}
