% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.evalMto1}
\alias{.evalMto1}
\title{Evaluate M-to-1 Orthologs for Splits}
\usage{
.evalMto1(ortho, gff = gff)
}
\description{
This function evaluates M-to-1 orthologs to identify cases where a M-to-1 relationship might actually be a result of gene splits.
}
