% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.evalMtoM}
\alias{.evalMtoM}
\title{Evaluate M-to-M Orthologs for Splits}
\usage{
.evalMtoM(ortho, gff = gff)
}
\description{
This function evaluates M-to-M orthologs to identify cases where an M-to-M relationship might actually be a result of gene splits.
}
