% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.filterOG}
\alias{.filterOG}
\title{Filter Ortholog Groups}
\usage{
.filterOG(target, ortho, pident, evalue, qcovs)
}
\description{
This function filters ortholog groups based on identity percentage, e-value, and query coverage.
}
