% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.find1to1gene}
\alias{.find1to1gene}
\title{Find 1-to-1 Gene Pairs from M-to-M Classes}
\usage{
.find1to1gene(target)
}
\description{
This function identifies the best 1-to-1 gene pairs from the M-to-M ortholog class by selecting pairs with the lowest mutual_e value and the highest mutual_ci value.
}
