% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.find1to1split}
\alias{.find1to1split}
\title{Find 1-to-1 Gene Splits}
\usage{
.find1to1split(seqid, geneid, pair_id, gff)
}
\description{
This function identifies cases where an M-to-M relationship might actually be a result of gene splits.
}
