% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.findBestPair}
\alias{.findBestPair}
\title{Find the Best Pair of Orthologous Genes}
\usage{
.findBestPair(object, ortho)
}
\description{
This function identifies the best pair of orthologous genes based on several criteria, including RBBH status, mutual e-value, mutual CI, and sequence length.
}
