% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.findNearestAnchor}
\alias{.findNearestAnchor}
\title{Find Nearest Anchor}
\usage{
.findNearestAnchor(gff, anchor, omit_chr)
}
\description{
This function finds the nearest anchor for each gene in the GFF data.
}
