% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.fixGFF}
\alias{.fixGFF}
\title{Fix and Standardize GFF Annotations}
\usage{
.fixGFF(gff)
}
\description{
This function applies a series of fixes and standardizations to GFF annotations.
}
