% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.fixGFFrange}
\alias{.fixGFFrange}
\title{Fix GFF Range}
\usage{
.fixGFFrange(gff)

.fixGFFrange(gff)
}
\description{
This function adjusts the start and end positions of transcripts and genes to cover the complete range of their member elements.

This function adjusts the start and end positions of transcripts and genes to cover the complete range of their member elements.
}
