% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.flipMinusStrand}
\alias{.flipMinusStrand}
\title{Flip start and end positions for minus strand gaps}
\usage{
.flipMinusStrand(df)
}
\arguments{
\item{df}{A data.frame containing gap information.}
}
\value{
A data.frame with flipped start and end positions for minus strand gaps.
}
\description{
This function flips the start and end positions for gaps on the minus strand.
}
