% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.gapGR}
\alias{.gapGR}
\title{Create GRanges object for genomic gaps}
\usage{
.gapGR(gr, chrLen)
}
\arguments{
\item{gr}{A GRanges object containing LCBs.}

\item{chrLen}{A data.frame containing chromosome lengths.}
}
\value{
A GRanges object representing the gaps between LCBs.
}
\description{
This function creates a GRanges object for genomic gaps between Locally Collinear Blocks (LCBs).
}
