% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.gapGrEndEdge}
\alias{.gapGrEndEdge}
\title{Create GRanges for end edges of gaps}
\usage{
.gapGrEndEdge(gr, gap_gr, chrLen, n_gr)
}
\arguments{
\item{gr}{A GRanges object containing LCBs.}

\item{gap_gr}{A data.frame containing gap information.}

\item{chrLen}{A data.frame containing chromosome lengths.}

\item{n_gr}{The number of GRanges in the input.}
}
\value{
A data.frame with updated gap information for end edges.
}
\description{
This function creates a GRanges object for the gaps at the end edges of chromosomes.
}
