% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.gapGrStartEdge}
\alias{.gapGrStartEdge}
\title{Create GRanges for start edges of gaps}
\usage{
.gapGrStartEdge(gr, gap_gr)
}
\arguments{
\item{gr}{A GRanges object containing LCBs.}

\item{gap_gr}{A data.frame containing gap information.}
}
\value{
A data.frame with updated gap information for start edges.
}
\description{
This function creates a GRanges object for the gaps at the start edges of chromosomes.
}
