% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.gene2block}
\alias{.gene2block}
\title{Map genes to blocks}
\usage{
.gene2block(gr, rbbh_id, gff)
}
\description{
This function maps genes to their corresponding blocks (LCBs or CBIs) based on overlaps.
}
