% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{.genomeSummary}
\alias{.genomeSummary}
\title{Summarize genome information}
\usage{
.genomeSummary(genome)
}
\arguments{
\item{genome}{Path to the genome file.}
}
\value{
A list containing genome names and lengths.
}
\description{
This function reads a genome file and returns a summary containing the names and lengths of the sequences.
}
