% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.get1to1}
\alias{.get1to1}
\title{Get 1-to-1 LCB pairs}
\usage{
.get1to1(h5, valid)
}
\arguments{
\item{h5}{An HDF5 file handle.}

\item{valid}{A logical vector indicating valid LCBs.}
}
\value{
A data.frame containing 1-to-1 LCB pairs.
}
\description{
This function retrieves 1-to-1 Locally Collinear Block (LCB) pairs from the HDF5 file.
}
