% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.getCDSblock}
\alias{.getCDSblock}
\title{Get CDS Blocks}
\usage{
.getCDSblock(gff)
}
\description{
This function extracts the concatenated CDS blocks for each transcript in the GFF data.
}
