% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.getCDSlen}
\alias{.getCDSlen}
\title{Calculate CDS Lengths}
\usage{
.getCDSlen(gff)
}
\description{
This function calculates the lengths of coding sequences (CDS) for each transcript in a GFF file.
}
