% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.getIDmap}
\alias{.getIDmap}
\title{Create ID Map for Gene Overlaps}
\usage{
.getIDmap(gff_gene, ol_list)
}
\description{
This function creates an ID map for overlapping genes by selecting a representative gene ID for each set of overlapping genes.
}
