% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getInterval}
\alias{.getInterval}
\title{Get intervals for blocks}
\usage{
.getInterval(g2b, gr)
}
\description{
This function gets intervals for blocks, considering non-NA values and creating interval strings.
}
