% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getMetrics}
\alias{.getMetrics}
\title{Get Metrics for Ortholog Prediction}
\usage{
.getMetrics(rbh_dist, tp_list, fp_list)
}
\description{
This function calculates various metrics (precision, recall, specificity, accuracy, and F-measure)
for ortholog prediction based on the distances to anchors.
}
