% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getMutualScore}
\alias{.getMutualScore}
\title{Calculate Mutual Scores for Orthologs}
\usage{
.getMutualScore(ortho)
}
\description{
This function calculates mutual coverage identity (CI) and mutual e-value for ortholog pairs.
}
