% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getNearestAnchorIndex}
\alias{.getNearestAnchorIndex}
\title{Get Nearest Anchor Index}
\usage{
.getNearestAnchorIndex(index, gff)

.getNearestAnchorIndex(index, gff)
}
\description{
This function finds the nearest anchor index for a given gene.

This function finds the nearest anchor index for a given gene.
}
