% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getOrphanGene}
\alias{.getOrphanGene}
\title{Identify Orphan Genes}
\usage{
.getOrphanGene(ortho, tx_orphan)
}
\description{
This function identifies genes that do not have orthologous pairs.
}
