% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{.getRBH}
\alias{.getRBH}
\title{Find Reciprocal Best Hits (RBH)}
\usage{
.getRBH(df1, df2)
}
\arguments{
\item{df1}{A data.frame containing BLAST search results from query to subject.}

\item{df2}{A data.frame containing BLAST search results from subject to query.}
}
\value{
A data.frame containing the reciprocal best hits with relevant BLAST statistics.
}
\description{
This function identifies Reciprocal Best Hits (RBH) between two sets of BLAST search results.
}
