% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_reorgOrthopairs.R
\name{.getRenameList}
\alias{.getRenameList}
\title{Build split-gene rename list and per-genome gene sets from orthopair TSVs}
\usage{
.getRenameList(orthopair_files, rename = FALSE, n_threads = 1L, verbose = TRUE)
}
\description{
Build split-gene rename list and per-genome gene sets from orthopair TSVs
}
