% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.getUniqTx}
\alias{.getUniqTx}
\title{Get Unique Transcripts}
\usage{
.getUniqTx(gff1, gff1_block_uniq)
}
\description{
This function extracts unique transcripts and their associated elements (e.g., CDS) from a GFF object.
}
