% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeAnchorList}
\alias{.makeAnchorList}
\title{Make Anchor List}
\usage{
.makeAnchorList(gff_ls, h5, omit_chr)
}
\description{
This function creates a list of nearest anchors for query and subject genomes.
}
