% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.makeCDS}
\alias{.makeCDS}
\title{Create CDS sequences from GFF and genome files}
\usage{
.makeCDS(gff, genome)
}
\arguments{
\item{gff}{Path to the GFF file.}

\item{genome}{Path to the genome file.}
}
\value{
A DNAStringSet object containing CDS sequences.
}
\description{
This function generates CDS sequences from provided GFF and genome files.
}
