% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{.makeFASTA}
\alias{.makeFASTA}
\title{Merge CDS files into a single FASTA file}
\usage{
.makeFASTA(cds_fn)
}
\arguments{
\item{cds_fn}{A character vector containing paths to CDS files.}
}
\value{
The path to the merged FASTA file.
}
\description{
This function reads multiple CDS files and merges them into a single FASTA file.
}
