% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeGRanges}
\alias{.makeGRanges}
\title{Convert LCB pairs to GRanges object}
\usage{
.makeGRanges(df, genome)
}
\arguments{
\item{df}{A data.frame containing LCB pairs.}

\item{genome}{A character string indicating whether to create GRanges for "query" or "subject".}
}
\value{
A GRanges object for the specified genome.
}
\description{
This function converts Locally Collinear Block (LCB) pairs to a GRanges object for either the query or subject genome.
}
