% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeObject}
\alias{.makeObject}
\title{Create a working object for ortholog filtering}
\usage{
.makeObject(object, h5)
}
\arguments{
\item{object}{A SynogDB object.}

\item{h5}{An HDF5 file handle.}
}
\value{
A list containing the working object data.
}
\description{
This function creates a working object containing necessary data for ortholog filtering based on LCBs.
}
