% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeOutput}
\alias{.makeOutput}
\title{Generate Output for Syntenic Ortholog Pairs}
\usage{
.makeOutput(syn_og, orphan)
}
\description{
This function generates the output for syntenic ortholog pairs, including a summary and orphan genes.
}
