% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeSummary}
\alias{.makeSummary}
\title{Generate Summary of Orthologous and Orphan Genes}
\usage{
.makeSummary(ortho, orphan)
}
\description{
This function creates a summary data frame that provides counts of orthologous gene pairs classified by type and counts of orphan genes.
}
