% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.makeSynogDF}
\alias{.makeSynogDF}
\title{Create Syntenic Ortholog Data Frame}
\usage{
.makeSynogDF(rbh, h5)
}
\description{
This function creates a data frame of syntenic orthologs based on RBH information.
}
