% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.mapEngine}
\alias{.mapEngine}
\title{Execute Protein Mapping Between Genomes}
\usage{
.mapEngine(
  object,
  subject_prot,
  query_prot,
  out_prefix,
  miniprot_bin,
  conda,
  condaenv,
  n_core,
  overlap,
  len_diff
)
}
\description{
This function maps proteins between the query and subject genomes using the specified miniprot binary and number of cores. It performs both directions of mapping and organizes the results.
}
