% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.mergeGFF}
\alias{.mergeGFF}
\title{Merge GFF Annotations}
\usage{
.mergeGFF(gff1, gff2 = NULL, ann_priority)
}
\description{
This function merges GFF annotations by combining two GFF files, removing duplicate gene and transcript entries, and creating a unified annotation.
}
