% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.mergeSynogDF}
\alias{.mergeSynogDF}
\title{Merge Syntenic Ortholog Data Frames}
\usage{
.mergeSynogDF(syn_og, rbbh_og, gff_ls)
}
\description{
This function merges syntenic ortholog data frames with RBBH orthologs.
}
