% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.numberingOG}
\alias{.numberingOG}
\title{Number Ortholog Groups}
\usage{
.numberingOG(ortho)
}
\description{
This function assigns unique identifiers to ortholog groups (OGs) based on the ortholog relationships between query and subject sequences.
}
