% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.numberingOGgene}
\alias{.numberingOGgene}
\title{Number Orthologous Groups (OGs) at the Gene Level}
\usage{
.numberingOGgene(ortho)
}
\description{
This function assigns a unique Orthologous Group (OG) number to each gene pair based on their ortholog classification (1-to-1, 1-to-M, M-to-1, or M-to-M).
}
