% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.order}
\alias{.order}
\title{Order LCB pairs}
\usage{
.order(df, by = "query")
}
\arguments{
\item{df}{A data.frame containing LCB pairs.}

\item{by}{A character string indicating whether to order by "query" or "subject" (default is "query").}
}
\value{
An ordered data.frame of LCB pairs.
}
\description{
This function orders Locally Collinear Block (LCB) pairs based on query or subject genome coordinates.
}
