% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orderGFF}
\alias{.orderGFF}
\title{Order GFF data}
\usage{
.orderGFF(gff)
}
\arguments{
\item{gff}{A GRanges object containing the GFF data.}
}
\value{
An ordered GRanges object.
}
\description{
This function orders GFF entries based on chromosome and start position.
}
