% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orgAnchor}
\alias{.orgAnchor}
\title{Organize Anchor Data}
\usage{
.orgAnchor(anchor, query_anchor, subject_anchor)
}
\description{
This function organizes anchor data for query and subject genomes.
}
