% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.orgCDS}
\alias{.orgCDS}
\title{Organize CDS Elements}
\usage{
.orgCDS(gff1, gff_tx)
}
\description{
This function organizes CDS elements in a GFF file by assigning gene IDs and updating parent IDs.
}
