% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.orgGene}
\alias{.orgGene}
\title{Organize Gene Elements}
\usage{
.orgGene(gff_tx)
}
\description{
This function organizes gene elements in a GFF file by creating gene entries from the organized transcript data.
}
