% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.orgGFF}
\alias{.orgGFF}
\title{Organize GFF Data}
\usage{
.orgGFF(gff1, gff2, gff3, prefix, overlap = FALSE, len_diff)
}
\description{
This function organizes and validates GFF data, filtering transcripts and organizing genes, transcripts, and CDS features.
}
