% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.orgTX}
\alias{.orgTX}
\title{Organize Transcripts}
\usage{
.orgTX(gff, prefix)
}
\description{
This function organizes transcripts in a GFF file by assigning gene IDs and updating transcript IDs.
}
