% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orthoIn1to1cbi}
\alias{.orthoIn1to1cbi}
\title{Filter orthologs in 1-to-1 CBI}
\usage{
.orthoIn1to1cbi(obj)
}
\arguments{
\item{obj}{A working object containing necessary data.}
}
\value{
A working object with filtered orthologs.
}
\description{
This function filters ortholog pairs that are in 1-to-1 Chromosome Break Intervals (CBI).
}
