% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orthoIn1to1lcb}
\alias{.orthoIn1to1lcb}
\title{Filter orthologs in 1-to-1 LCBs}
\usage{
.orthoIn1to1lcb(obj)
}
\arguments{
\item{obj}{A working object containing necessary data.}
}
\value{
A working object with filtered orthologs.
}
\description{
This function filters ortholog pairs that are in 1-to-1 Locally Collinear Blocks (LCBs).
}
