% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orthoInBlock}
\alias{.orthoInBlock}
\title{Filter ortholog pairs in blocks}
\usage{
.orthoInBlock(
  gr_query,
  gr_subject,
  rbbh,
  gff_query,
  gff_subject,
  interval = FALSE
)
}
\arguments{
\item{gr_query}{A GRanges object for query genome.}

\item{gr_subject}{A GRanges object for subject genome.}

\item{rbbh}{A data.frame containing reciprocal best BLAST hits.}

\item{gff_query}{A GRanges object containing query GFF data.}

\item{gff_subject}{A GRanges object containing subject GFF data.}

\item{interval}{Logical indicating whether to use intervals (default is FALSE).}
}
\value{
A data.frame containing the filtered ortholog pairs.
}
\description{
This function filters ortholog pairs that are within specified blocks.
}
