% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.orthoInNon1to1lcb}
\alias{.orthoInNon1to1lcb}
\title{Filter orthologs in non-1-to-1 LCBs}
\usage{
.orthoInNon1to1lcb(obj)
}
\arguments{
\item{obj}{A working object containing necessary data.}
}
\value{
A working object with filtered orthologs.
}
\description{
This function filters ortholog pairs that are in non-1-to-1 Locally Collinear Blocks (LCBs).
}
