% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.phaseMinus}
\alias{.phaseMinus}
\title{Adjust Phase for CDS Features on Minus Strand}
\usage{
.phaseMinus(gff_cds)
}
\description{
This function adjusts the phase for CDS features on the minus strand in GFF annotations.
}
