% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.phasePlus}
\alias{.phasePlus}
\title{Adjust Phase for CDS Features on Plus Strand}
\usage{
.phasePlus(gff_cds)
}
\description{
This function adjusts the phase for CDS features on the plus strand in GFF annotations.
}
