% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.pick1to1rest}
\alias{.pick1to1rest}
\title{Pick 1-to-1 Gene Relationships from the Remaining Pairs}
\usage{
.pick1to1rest(rest, pair_id)
}
\description{
This function identifies 1-to-1 gene relationships from the remaining pairs by removing pairs where either gene is duplicated.
}
