% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.prepGenome}
\alias{.prepGenome}
\title{Prepare genome for SibeliaZ}
\usage{
.prepGenome(genome, label)
}
\arguments{
\item{genome}{Path to the genome file.}

\item{label}{Label to prepend to sequence names.}
}
\value{
Path to the temporary FASTA file.
}
\description{
This function reads a genome file and modifies sequence names with a given label,
then writes the modified sequences to a temporary FASTA file.
}
