% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.prepGFF}
\alias{.prepGFF}
\title{Prepare GFF Data for Gene Processing}
\usage{
.prepGFF(object)
}
\description{
This function imports and filters GFF data to extract CDS features and ensures that the Parent attributes are properly unlisted.
}
