% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_orthology.R
\name{.process_one_pair_dataframe}
\alias{.process_one_pair_dataframe}
\title{Process one genome pair RBH file using R/05_functions_orthology.R algorithm}
\usage{
.process_one_pair_dataframe(
  rbh_fn,
  pair_id,
  gff_data_all,
  gff_lookup,
  genome_width,
  pident_threshold_quantile,
  mutual_ci_min,
  load_all_gff,
  verbose
)
}
\description{
Process one genome pair RBH file using R/05_functions_orthology.R algorithm
}
