% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_reorgOrthopairs.R
\name{.reorgOrthoPair}
\alias{.reorgOrthoPair}
\title{Reorganise orthopair TSVs into per-pair tables suitable for graph building}
\usage{
.reorgOrthoPair(
  orthopair_files,
  reorg_list,
  n_threads = 1L,
  rename = FALSE,
  verbose = TRUE
)
}
\description{
Reorganise orthopair TSVs into per-pair tables suitable for graph building
}
