% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.replaceIDs}
\alias{.replaceIDs}
\title{Replace Gene IDs with Transcript IDs for Split Gene Relationships}
\usage{
.replaceIDs(ortho)
}
\description{
This function updates the gene IDs in the ortholog pairs with their respective transcript IDs for split gene relationships.
}
