% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.setAnchor2RBH}
\alias{.setAnchor2RBH}
\title{Set Anchor Information to RBH}
\usage{
.setAnchor2RBH(rbh, anchor_ls)
}
\description{
This function sets the anchor information to the reciprocal best hits (RBH).
}
