% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.setGeneID}
\alias{.setGeneID}
\title{Set Gene IDs for GFF Annotations}
\usage{
.setGeneID(gff)
}
\description{
This function sets the gene_id attribute for GFF annotations based on their parent-child relationships.
}
