% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.setGeneID2Anchor}
\alias{.setGeneID2Anchor}
\title{Set Gene IDs to Anchors}
\usage{
.setGeneID2Anchor(anchor, gff_ls)
}
\description{
This function sets the gene IDs for query and subject anchors.
}
