% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.setGeneIDsynog}
\alias{.setGeneIDsynog}
\title{Set Gene IDs for Syntenic Orthologs}
\usage{
.setGeneIDsynog(rbbh_og, gff_ls)
}
\description{
This function sets the gene IDs for query and subject sequences in the syntenic orthologs data frame.
}
