% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.solveMultiProjectionQ2S}
\alias{.solveMultiProjectionQ2S}
\title{Solve Multiple Projections from Query to Subject}
\usage{
.solveMultiProjectionQ2S(index, rbh, anchor_ls)
}
\description{
This function solves multiple projections from query to subject anchors.
}
