% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.solveMultiProjectionS2Q}
\alias{.solveMultiProjectionS2Q}
\title{Solve Multiple Projections from Subject to Query}
\usage{
.solveMultiProjectionS2Q(index, rbh, anchor_ls)
}
\description{
This function solves multiple projections from subject to query anchors.
}
