% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{.sortLabels}
\alias{.sortLabels}
\title{Sort labels}
\usage{
.sortLabels(x, rev = FALSE)
}
\arguments{
\item{x}{A vector of labels to sort.}

\item{rev}{Logical, whether to reverse the sort order (default is FALSE).}
}
\value{
A factor with sorted labels.
}
\description{
This function sorts labels, optionally in reverse order.
}
