% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.subsetGFF}
\alias{.subsetGFF}
\title{Subset and order GFF files}
\usage{
.subsetGFF(query_gff, subject_gff)
}
\arguments{
\item{query_gff}{A GRanges object containing the query GFF data.}

\item{subject_gff}{A GRanges object containing the subject GFF data.}
}
\value{
A list containing the subsetted and ordered query and subject GFF data.
}
\description{
This function subsets GFF files to include only transcripts and mRNA, and orders them.
}
