% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.traceAnchorQ2S}
\alias{.traceAnchorQ2S}
\title{Trace Anchor Distance from Query to Subject}
\usage{
.traceAnchorQ2S(rbh, anchor_ls)
}
\description{
This function calculates the distance from query anchors to subject anchors.
}
