% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.traceAnchorS2Q}
\alias{.traceAnchorS2Q}
\title{Trace Anchor Distance from Subject to Query}
\usage{
.traceAnchorS2Q(rbh, anchor_ls)
}
\description{
This function calculates the distance from subject anchors to query anchors.
}
