% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.validTX}
\alias{.validTX}
\title{Validate Transcripts}
\usage{
.validTX(gff1, gff2, gff3, overlap, len_diff)
}
\description{
This function validates transcripts by checking overlaps and length differences with reference GFF data.
}
