% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.writeProtFASTA}
\alias{.writeProtFASTA}
\title{Write Protein Sequences to a FASTA File}
\usage{
.writeProtFASTA(id, prot_fn, gff_fn, fn)

.writeProtFASTA(id, prot_fn, gff_fn, fn)
}
\description{
This function writes protein sequences to a FASTA file based on the specified IDs.

This function writes protein sequences to a FASTA file based on the specified IDs.
}
