% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{geneOrtho}
\alias{geneOrtho}
\title{Identify Gene Orthologs}
\usage{
geneOrtho(object)
}
\arguments{
\item{object}{A SynogDB object.}
}
\description{
This function identifies gene orthologs based on syntenic ortholog pairs and assigns ortholog groups.
}
