% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{getLCBpairs}
\alias{getLCBpairs}
\title{Get pairs of blocks between query and subject genomes}
\usage{
getLCBpairs(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\description{
This function generates a table showing all pairs of blocks between the query and subject genomes.
}
