% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{getMeta}
\alias{getMeta}
\title{Get Meta data of pairing}
\usage{
getMeta(object = NULL, hdf5_fn = NULL)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{hdf5_fn}{A path to a hdf5 file storing OrthoPaiR results.}
}
\value{
A dataframe containing ortholog pairs.
}
\description{
This function retrieves meta data from the OrthoPairDB object.
}
