% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{lcbClassify}
\alias{lcbClassify}
\title{Classify LCBs}
\usage{
lcbClassify(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\description{
This function classifies Locally Collinear Blocks (LCBs) based on their genome origin (query or subject).
}
