% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{makeSynogDB}
\alias{makeSynogDB}
\title{Create a SynogDB object}
\usage{
makeSynogDB(
  query_genome,
  subject_genome,
  query_gff,
  subject_gff,
  query_cds,
  subject_cds,
  query_prot,
  subject_prot,
  hdf5_path = "./synog.h5",
  overwrite = FALSE
)
}
\arguments{
\item{query_genome}{Path to the query genome file.}

\item{subject_genome}{Path to the subject genome file.}

\item{query_gff}{Path to the query GFF file.}

\item{subject_gff}{Path to the subject GFF file.}

\item{query_cds}{Path to the query CDS file.}

\item{subject_cds}{Path to the subject CDS file.}

\item{query_prot}{Path to the query protein file.}

\item{subject_prot}{Path to the subject protein file.}

\item{hdf5_path}{Path to the HDF5 file (default is "./synog.h5").}
}
\value{
A SynogDB object.
}
\description{
This function creates a SynogDB object containing genome, GFF, CDS, and protein information
for query and subject genomes. It also initializes an HDF5 file to store related data.
}
