% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{mapProt}
\alias{mapProt}
\title{Map Proteins Between Genomes}
\usage{
mapProt(
  object,
  out_dir,
  miniprot_bin = "miniprot",
  conda = "conda",
  condaenv = NULL,
  n_threads,
  len_diff = 0.5,
  overlap = TRUE
)
}
\arguments{
\item{object}{A SynogDB object.}

\item{miniprot_bin}{A character string specifying the path to the miniprot binary.}

\item{n_threads}{An integer specifying the number of cores to use for the mapping.}

\item{len_diff}{A numeric value specifying the maximum allowable length difference for proteins. Default is 0.5.}

\item{out_prefix}{A character string specifying the prefix for the output files.}
}
\value{
None. The function performs the mapping and writes the results to the output files specified by \code{out_prefix}.
}
\description{
This function maps proteins between the query and subject genomes using the specified miniprot binary and number of cores.
}
