% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{plotLCBpairs}
\alias{plotLCBpairs}
\title{Plot LCB pairs}
\usage{
plotLCBpairs(
  object,
  class = "both",
  size = 0.5,
  color_1to1 = "darkgreen",
  color_non1to1 = "blue",
  chr_border_color = "gray60"
)
}
\arguments{
\item{object}{A SynogDB object.}

\item{class}{The class of LCB pairs to plot ("1to1", "non1to1", or "both").}

\item{size}{Point size for the plot (default is 0.5).}

\item{color_1to1}{Color for 1-to-1 pairs (default is "darkgreen").}

\item{color_non1to1}{Color for non-1-to-1 pairs (default is "blue").}

\item{chr_border_color}{Color for chromosome borders (default is "gray60").}
}
\value{
A ggplot object.
}
\description{
This function plots pairs of blocks between the query genome and the subject genome.
}
