% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_rbh.R
\name{rbh}
\alias{rbh}
\title{Create BLAST databases and perform reciprocal BLAST}
\usage{
rbh(
  working_dir,
  blast_path,
  target_pair = NULL,
  n_threads = 1,
  overwrite = FALSE
)
}
\description{
This function prepare BLAST databases and perform BLAST searches reciprocally
on given input datasets to create the list of reciprocal BLAST hits.
}
