% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions_reformatFiles.R
\name{reformatFiles}
\alias{reformatFiles}
\title{Make input object and validate working directory}
\usage{
reformatFiles(object, working_dir, overwrite, n_threads)
}
\description{
This function organize input files and structure working directory with
confirming previous outputs
}
