% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_functions_pipeline.R
\name{runOrthoPair}
\alias{runOrthoPair}
\title{Function to go through the OrthoPair pipeline}
\usage{
runOrthoPair(
  query_genome,
  subject_genome,
  query_gff,
  subject_gff,
  query_cds,
  subject_cds,
  query_prot,
  subject_prot,
  hdf5_path = "./orthopair.h5",
  sibeliaz_out_dir = "./sibeliaz_out",
  sibeliaz_bin = "sibeliaz",
  maf2synteny_bin = "maf2synteny",
  conda = "conda",
  sibeliaz_condaenv = "sibeliaz",
  miniprot_bin = "miniprot",
  miniprot_condaenv = "miniprot",
  miniprot_out_dir = "./miniprot_out/",
  n_threads = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\description{
This is a wrapper function to execute the series of functions to go through
the OrthoPair pipeline.
}
