% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{runSibeliaZ}
\alias{runSibeliaZ}
\title{Run SibeliaZ}
\usage{
runSibeliaZ(
  object,
  out_dir,
  sibeliaz_bin = "sibeliaz",
  maf2synteny_bin = "maf2synteny",
  conda = "conda",
  condaenv = NULL,
  run_sibeliaz = TRUE
)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{out_dir}{Output directory for SibeliaZ results.}

\item{sibeliaz_bin}{Path to the SibeliaZ binary (default is "sibeliaz").}

\item{maf2synteny_bin}{Path to the maf2synteny binary (default is "maf2synteny").}

\item{conda}{Path to the conda executable (default is "conda").}

\item{condaenv}{Name of the conda environment to use (default is NULL).}

\item{run_sibeliaz}{Logical, whether to run SibeliaZ (default is TRUE).}
}
\description{
This function runs SibeliaZ to detect synteny blocks between the query and subject genomes
and processes the output to store in the OrthoPairDB object.
}
