% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{showLCB}
\alias{showLCB}
\title{Show LCB information}
\usage{
showLCB(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\description{
This function displays information about the Locally Collinear Blocks (LCBs) in the OrthoPairDB object.
}
