% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{sibeliaLCB2DF}
\alias{sibeliaLCB2DF}
\title{Convert SibeliaZ output to a data.frame}
\usage{
sibeliaLCB2DF(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\value{
A data.frame containing LCB information.
}
\description{
This function processes the output of SibeliaZ and converts it into a data.frame
containing information about Locally Collinear Blocks (LCBs).
}
