% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{sibeliaRAW2DF}
\alias{sibeliaRAW2DF}
\title{Convert SibeliaZ raw output to a data.frame}
\usage{
sibeliaRAW2DF(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\value{
A data.frame containing raw LCB information.
}
\description{
This function processes the raw output of SibeliaZ and converts it into a data.frame
containing information about Locally Collinear Blocks (LCBs).
}
