% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{splitGenes}
\alias{splitGenes}
\title{Split Genes Based on Orthologous Relationships}
\usage{
splitGenes(object)
}
\arguments{
\item{object}{A SynogDB object containing necessary HDF5 and GFF files.}
}
\description{
This function processes genewise ortholog pairs to identify and handle split genes. It evaluates the genewise classes, replaces gene IDs, and generates orthologous gene groups.
}
