% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions_synteny.R
\name{statsLCB}
\alias{statsLCB}
\title{Summarize LCB statistics}
\usage{
statsLCB(object)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\description{
This function calculates and summarizes statistics for LCBs (Locally Collinear Blocks) in the OrthoPairDB object.
}
