% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{syntenyOrtho}
\alias{syntenyOrtho}
\title{Filter ortholog pairs based on gene synteny}
\usage{
syntenyOrtho(object, omit_chr = "", pident = 90, evalue = 1e-50, qcovs = 50)
}
\arguments{
\item{object}{A SynogDB object.}

\item{omit_chr}{A character string specifying chromosomes to omit (default is "").}

\item{pident}{Minimum percentage identity for BLAST hits (default is 90).}

\item{evalue}{Maximum e-value for BLAST hits (default is 1e-50).}

\item{qcovs}{Minimum query coverage for BLAST hits (default is 50).}
}
\description{
This function filters ortholog pairs based on gene synteny, using specified thresholds for percentage identity, e-value, and query coverage.
}
