% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{updateFiles}
\alias{updateFiles}
\title{Update SynogDB files}
\usage{
updateFiles(object)
}
\arguments{
\item{object}{A SynogDB object.}
}
\value{
The updated SynogDB object.
}
\description{
This function updates the GFF and CDS files in a SynogDB object from the HDF5 file.
}
