% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{.blast_filter}
\alias{.blast_filter}
\title{Filter BLAST search results}
\usage{
.blast_filter(rbh_out)
}
\arguments{
\item{blast_out}{A data.frame containing the BLAST search results.}
}
\value{
A filtered data.frame of BLAST search results.
}
\description{
This function filters BLAST search results based on specified criteria for percentage identity,
query coverage, and e-value. It can also identify the best hits for each query sequence.
}
