% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions_genemodeling.R
\name{.createFASTA}
\alias{.createFASTA}
\title{Create FASTA files}
\usage{
.createFASTA(object, out_dir, merge)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{out_dir}{Output directory for FASTA files.}
}
\description{
This function generates FASTA files for query and subject CDS from the OrthoPairDB object.
}
