% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{.getDist}
\alias{.getDist}
\title{Get Distance between Anchors and RBH}
\usage{
.getDist(gff_ls, h5, anchor_ls)
}
\description{
This function calculates the distance between anchors and reciprocal best hits (RBH).
}
