% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{getOrphan}
\alias{getOrphan}
\title{Get orphan information}
\usage{
getOrphan(object = NULL, hdf5_fn = NULL, loc = FALSE)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\value{
A dataframe containing orphan genes or transcripts.
}
\description{
This function retrieves orphan genes or transcripts information from the OrthoPairDB object.
}
