% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{getOrphan}
\alias{getOrphan}
\title{Get orphan information}
\usage{
getOrphan(object = NULL, h5_fn = NULL, gene = FALSE)
}
\arguments{
\item{object}{A SynogDB object.}

\item{gene}{Logical, whether to get orphan gene information.
Orphan transcript information will be returned if FALSE. (default is FALSE).}

\item{split}{Logical, whether to obtain split orphan gene information (default is FALSE).}
}
\value{
A dataframe containing orphan genes or transcripts.
}
\description{
This function retrieves orphan genes or transcripts information from the SynogDB object.
}
