% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{getOrthoPair}
\alias{getOrthoPair}
\title{Get OrthoPair information}
\usage{
getOrthoPair(object = NULL, h5_fn = NULL, gene = FALSE)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{h5_fn}{A path to a hdf5 file storing OrthoPaiR results.}

\item{gene}{Logical, whether to get genewise orthology information (default is FALSE).}

\item{split}{Logical, whether to get split genewise orthology information (default is FALSE).}
}
\value{
A dataframe containing ortholog pairs.
}
\description{
This function retrieves ortholog pairs information from the OrthoPairDB object.
}
