% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_functions_pipeline.R
\name{orthopair}
\alias{orthopair}
\title{Function to go through the OrthoPair pipeline for all combinations of genomes}
\usage{
orthopair(
  in_list,
  hdf5_out_dir = "./hdf5",
  sibeliaz_out_dir = "./sibeliaz_out",
  sibeliaz_bin = "sibeliaz",
  maf2synteny_bin = "maf2synteny",
  conda = "conda",
  sibeliaz_condaenv = "sibeliaz",
  miniprot_bin = "miniprot",
  miniprot_condaenv = "miniprot",
  miniprot_out_dir = "./miniprot_out/",
  n_threads = NULL,
  one_to_many = FALSE,
  verbose = TRUE,
  overwrite = FALSE,
  resume = FALSE,
  redo = NULL,
  param_list = NULL
)
}
\description{
This is a wrapper function to execute the series of functions to go through
the OrthoPair pipeline for all combinations of genomes.
}
