% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_functions_pipeline.R
\name{orthopair}
\alias{orthopair}
\title{Function to go through the OrthoPair pipeline for all combinations of genomes}
\usage{
orthopair(
  in_list,
  working_dir = ".",
  conda = "conda",
  miniprot_bin = "miniprot",
  miniprot_condaenv = "miniprot",
  diamond_exec_path = NULL,
  n_threads = NULL,
  target_pair = NULL,
  use_prot = FALSE,
  verbose = TRUE,
  overwrite = FALSE,
  resume = FALSE,
  redo = NULL,
  param_list = NULL,
  orthopair = TRUE,
  reorg = TRUE,
  makegraph = TRUE,
  output_table = FALSE
)
}
\description{
This is a wrapper function to execute the series of functions to go through
the OrthoPair pipeline for all combinations of genomes.
}
