% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_functions_pipeline.R
\name{orthopair}
\alias{orthopair}
\title{Function to go through the OrthoPair pipeline for all combinations of genomes}
\usage{
orthopair(
  in_list,
  working_dir = ".",
  miniprot_path = "",
  blast_path = "",
  diamond_path = "",
  target_pair = NULL,
  use_prot = FALSE,
  run_miniprot = FALSE,
  orthopair = TRUE,
  reorg = TRUE,
  rename = TRUE,
  n_threads = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\description{
This is a wrapper function to execute the series of functions to go through
the OrthoPair pipeline for all combinations of genomes.
}
