% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functions_blast.R
\name{rbh}
\alias{rbh}
\title{Execute Reciprocal BLAST Hits (RBH) search}
\usage{
rbh(
  object,
  n_threads = 1,
  use_prot = FALSE,
  blast_path = "",
  diamond_path = ""
)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{n_threads}{Number of threads to use for BLAST (default is 1).}

\item{db1}{Path to the BLAST database for query genome (default is NULL).}

\item{db2}{Path to the BLAST database for subject genome (default is NULL).}
}
\description{
This function performs a Reciprocal BLAST Hits (RBH) search between query and subject genomes.
}
