% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{summaryOrthoPair}
\alias{summaryOrthoPair}
\title{Summarize OrthoPairDB object}
\usage{
summaryOrthoPair(object = NULL, h5_fn = NULL, gene = FALSE)
}
\arguments{
\item{object}{A OrthoPairDB object.}

\item{h5_fn}{A path to a hdf5 file storing OrthoPaiR results.}
}
\value{
A summary dataframe.
}
\description{
This function provides a summary of the OrthoPairDB object, optionally for gene or transcript information.
}
