% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions_IO.R
\name{summarySynog}
\alias{summarySynog}
\title{Summarize SynogDB object}
\usage{
summarySynog(object = NULL, h5_fn = NULL, gene = FALSE)
}
\arguments{
\item{object}{A SynogDB object.}

\item{h5_fn}{A path to a hdf5 file storing Synog results.}
}
\value{
A summary dataframe.
}
\description{
This function provides a summary of the SynogDB object, optionally for gene or transcript information.
}
