% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_functions_orthology.R
\name{syntenicOrtho}
\alias{syntenicOrtho}
\title{Filter ortholog pairs based on Locally Collinear Blocks (LCBs)}
\usage{
syntenicOrtho(object, rbbh_mci_threshold = 0.1, rbh_mci_threshold = 0.4)
}
\arguments{
\item{object}{A OrthoPairDB object.}
}
\description{
This function filters ortholog pairs based on their presence in Locally Collinear Blocks (LCBs).
}
